+++ Common/Network/PacketServer.cs
@@ -14,6 +14,7 @@ namespace Common.Network
         }
 
         public PacketServer(AuthCMD authOpcode) : this((byte)authOpcode) { }
+        public PacketServer(RealmEnums realmOpcode) : this((int)realmOpcode) { }
 
         public byte[] Packet => (BaseStream as MemoryStream)?.ToArray();